Lib

Description

This module contains all necessary functions to retrieve information from a Graph.

Some helper functions for getting the basic information from the Graph

vertexExist :: Vertex -> Graph -> Bool

Checks for the existence of a certain Vertex

verticesAmount :: Graph -> Int

Returns the amount of the vertices in the Graph

getAdjacentVertices :: Vertex -> Graph -> Maybe [Edge]

Gets a list of all the adjacent vertices with weights if they exist otherwise Nothing

getAllVertices :: Graph -> [Vertex]

Gets a list of all the existing Vertices

Functions to get information about the Edges

getEdgesAmount :: Graph -> Int

Get the total amount of Edges in the undirectional Graph

countEdges :: Vertex -> Graph -> Int

Counts the connected Edges of a certain Vertex

getWeightSum :: Vertex -> Graph -> Int

Get the total sum of all the Weights of the Edges from one Vertex

getAllEdges :: Vertex -> Graph -> [Edge]

Gets all the Edges from a certain Vertex

adjacentVerticesToString :: Vertex -> Graph -> [String]

Returns for a certain Vertex its edges with other vertices and weights

edgeToString :: Vertex -> [Edge] -> [String]

Puts info about the relation between a vertex and its edges in a list of Strings

Functions to get information about connections in the Graph

getOnlyConnVertices :: Vertex -> Graph -> [Vertex]

Gets all the adjacent connected Edges and returns all the Vertices of those Edges

twoVerticesConn :: Vertex -> Vertex -> Graph -> Bool

Checks if two Vertices are connected

getCommonVertices :: Vertex -> Vertex -> Graph -> [Vertex]

Returns any common connected Vertices between two Vertices

haveCommonVertices :: Vertex -> Vertex -> Graph -> Bool

Checks if two Vertices share a common connected Vertex

compareTwoLists :: [Vertex] -> [Vertex] -> [Vertex]

Compares two lists and returns the common Vertices